remove some currency name in public pledge payment

+++ src/views/accounting/chartOfAccount/Liabilities.vue
@@ -1,10 +1,10 @@
 <template>
-  <div class="row" v-for="(item, index) in data.accountHeadsDTO" :key="index">
+  <div class="container-fluid px-0" v-for="(item, index) in data.accountHeadsDTO" :key="index">
     <div class="col-md-12">
       <div class="row">
         <div class="col-12 py-2 mt-4 account-head">
           {{ item.name }} <small class="font-weight-normal">{{ item.groupSubHead }}</small
-          ><i class="pi pi-question-circle-o help" aria-hidden="true"></i>
+          ><el-icon :size="20"><QuestionFilled /></el-icon>
         </div>
       </div>
       <div
@@ -18,8 +18,8 @@
         </div>
         <div class="col-6 col-md-5">{{ itm.description }}</div>
         <div class="col-6 col-md-2 text-right">
-          <i class="pi pi-pencil c-pointer" aria-hidden="true" data-toggle="modal" data-target="#liabModal" @click="editAccount(item, itm)"></i>
-          <i class="pi pi-trash ml-3 c-pointer" aria-hidden="true"  @click="deleteAccount(itm.id, index, indx)"></i>
+          <el-icon data-toggle="modal" data-target="#liabModal" :size="20" @click="editAccount(item, itm)" class="c-pointer mx-2"><EditPen /></el-icon>
+          <el-icon class="c-pointer" :size="20" @click="deleteAccount(itm.id, index, indx)"><Delete /></el-icon>
         </div>
       </div>
       <div class="row row-border align-items-center py-3" v-if="item.accounts.length === 0">
@@ -35,7 +35,9 @@
               class="c-pointer text-decoration-none primary-text"
               data-toggle="modal"
               data-target="#liabModal"
-              ><i class="pi pi-plus-circle"></i>&nbsp; &nbsp; Add a new
+              >
+              <el-icon  :size="20"><CirclePlus /></el-icon>
+              &nbsp; &nbsp; Add a new
               Account</a
             >
           </div>
@@ -61,15 +63,13 @@
               <h5 class="modal-title font-weight-bold" id="exampleModalLabel">
                 Add an account
               </h5>
-              <button
-                type="button"
+              <el-button
                 class="close"
                 data-dismiss="modal"
                 aria-label="Close"
-                ref="closeAccountModalBtn"
               >
-                <span aria-hidden="true">&times;</span>
-              </button>
+                <span aria-hidden="true" class="mt-0" ref="closeAccountModalBtn"><el-icon :size="20"><Close /></el-icon></span>
+              </el-button>
             </div>
             <div class="modal-body">
               <CreateAccountModal
@@ -88,7 +88,6 @@
           </div>
         </div>
       </div>
-      <ConfirmDialog></ConfirmDialog>
       <!-- END BT -->
     </div>
   </div>
@@ -98,20 +97,15 @@
 import { computed, ref } from "vue";
 import CreateAccountModal from "./components/CreateAccountForm";
 import transactionals from "./utilities/transactionals";
-import ConfirmDialog from 'primevue/confirmdialog';
-import { useConfirm } from "primevue/useconfirm";
-import { useToast } from 'primevue/usetoast';
 import chart_of_accounts from '../../../services/financials/chart_of_accounts';
+import { ElMessage, ElMessageBox } from "element-plus";
 
 
 export default {
-    components: { CreateAccountModal, ConfirmDialog },
+    components: { CreateAccountModal },
     props: [ "data" ],
   setup(props, { emit }) {
     const view = ref(true);
-    const confirm = useConfirm();
-    const toast = useToast();
-
     const accounts = ref([]);
     const getAccounts = async () => {
       try {
@@ -162,29 +156,41 @@ export default {
     }
 
     const deleteAccount = (id, index, indx) => {
-      confirm.require({
-          message: 'Are you sure you want to delete this account?',
-          header: 'Confirmation',
-          icon: 'pi pi-exclamation-triangle',
-          acceptClass: 'confirm-delete',
-          rejectClass: 'cancel-delete',
-          accept: async () => {
-              //callback to execute when user confirms the action
-              try {
+      ElMessageBox.confirm(
+        "Are you sure you want to proceed?",
+        "Confirm delete",
+        {
+          confirmButtonText: "OK",
+          cancelButtonText: "Cancel",
+          type: "error",
+        }
+      )
+        .then( async () => {
+          try {
                 const response = await chart_of_accounts.deleteAccount(id);
-                toast.add({severity:'success', summary:'Account Deleted', detail: `${response.response}`, life: 3000});
-                emit("liability-deleted", index, indx);
+                ElMessage({
+                  type: "success",
+                  message: `${response.response}`,
+                  duration: 3000,
+                });
+               emit("liability-deleted", index, indx);
               } catch (error) {
-                toast.add({severity:'error', summary:'Delete Error', detail:'Account not deleted', life: 3000});
+                ElMessage({
+                  type: "error",
+                  message: "Account not deleted",
+                  duration: 3000,
+                });
                 console.log(error);
               }
-          },
-          reject: () => {
-            //callback to execute when user rejects the action
-            
-          }
-      });
-    }
+        })
+        .catch(() => {
+          ElMessage({
+            type: "info",
+            message: "Delete canceled",
+            duration: 5000,
+          });
+        });
+    };
 
     return {
       view,
